Move convert_pset_to_dict from ParticleFile to ParticleSet
